refactor(agentic): extract scraping persistence module and eliminate all CodeScene biomarkers (#87, #92)
