+++ backend/app/services/agentic/scraping_graph.py
@@ -4,29 +4,27 @@ from __future__ import annotations
 
 import json
 import logging
-import uuid
-from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any, TypedDict
 
 from langgraph.graph import END, START, StateGraph
-from sqlmodel import Session, select
 
-from app import crud
-from app.models import User
 from app.services.agentic.schemas import ScrapedBatchReport
+from app.services.agentic.scraping_persistence import (
+    _safe_int,
+    persist_scraped_batch_records,
+)
 from app.services.agentic.session_recovery_graph import (
     _detect_overlay,
     recover_page_session,
 )
-from app.services.agentic.tools.common import get_active_page, resolve_session
+from app.services.agentic.tools.common import get_active_page
 from app.services.browser.diagnostics import detect_page_state, extract_grok_summary
 from app.services.browser.manager import BrowserManager
 from scripts.scrape_trending_topics import (
     extract_topic_tweets,
     extract_trending_sidebar,
     navigate_to_trends,
-    parse_post_count,
 )
 
 logger = logging.getLogger(__name__)
@@ -94,48 +92,6 @@ def _parse_clamped_max_topics(val: Any, default: int = 3) -> int:
         return max(1, min(10, parsed))
     except (ValueError, TypeError):
         return default
-
-
-def _safe_int(val: Any, default: int = 0) -> int:
-    """Safely convert engagement metric or count to non-negative int."""
-    if val is None:
-        return default
-    if isinstance(val, int):
-        return max(0, val)
-    if isinstance(val, float):
-        return max(0, int(val))
-    if isinstance(val, str):
-        cleaned = val.strip().replace(",", "")
-        try:
-            return max(0, int(cleaned))
-        except (ValueError, TypeError):
-            parsed = parse_post_count(val)
-            return max(0, parsed if parsed is not None else default)
-    return default
-
-
-def _resolve_user_id(*, user_id: Any, session: Session) -> uuid.UUID:
-    """Resolve user ID string or UUID to a valid user UUID safely."""
-    if isinstance(user_id, uuid.UUID):
-        return user_id
-    if user_id:
-        try:
-            return uuid.UUID(str(user_id).strip())
-        except (ValueError, TypeError, AttributeError):
-            pass
-        try:
-            user = session.exec(select(User).where(User.email == str(user_id))).first()
-            if user:
-                return user.id
-        except Exception as e:
-            logger.warning(f"Error querying user by email/id {user_id}: {e}")
-    try:
-        first_user = session.exec(select(User)).first()
-        if first_user:
-            return first_user.id
-    except Exception as e:
-        logger.warning(f"Error querying fallback user: {e}")
-    return uuid.uuid4()
 
 
 def _verify_session_exists(*, user_id: str) -> tuple[bool, str | None]:
@@ -471,73 +427,17 @@ async def extract_topic_timelines_node(state: ScrapingGraphState) -> dict[str, A
     }
 
 
-def _build_topic_upsert_payload(
-    *,
-    resolved_user_id: uuid.UUID,
-    topic: dict[str, Any],
-    summaries: dict[str, str],
-    now: datetime,
-) -> tuple[str | None, dict[str, Any] | None]:
-    """Extract and sanitize database upsert dictionary for a trending topic."""
-    url = str(topic.get("topic_url") or topic.get("url", "")).strip()
-    if not url:
-        return None, None
-
-    title = str(topic.get("topic_title") or topic.get("title", "")).strip()
-    summary_val = summaries.get(url) or topic.get("summary")
-    post_count_val = _safe_int(topic.get("post_count"), default=0) or None
-
-    payload = {
-        "user_id": resolved_user_id,
-        "topic_url": url[:512],
-        "topic_title": title[:500],
-        "category": (
-            str(topic.get("category"))[:100]
-            if topic.get("category") is not None
-            else None
-        ),
-        "post_count": post_count_val,
-        "summary": str(summary_val) if summary_val is not None else None,
-        "last_seen_at": now,
-        "scraped_at": now,
-    }
-    return url, payload
-
-
-def _persist_single_topic_record(
-    *,
-    session: Session,
-    resolved_user_id: uuid.UUID,
-    topic: dict[str, Any],
-    tweets_map: dict[str, list[dict[str, Any]]],
-    summaries: dict[str, str],
-    now: datetime,
-) -> tuple[int, int]:
-    """Persist a single topic and its associated tweets to the database."""
-    url, topic_data = _build_topic_upsert_payload(
-        resolved_user_id=resolved_user_id,
-        topic=topic,
-        summaries=summaries,
-        now=now,
-    )
-    if not url or not topic_data:
-        return 0, 0
-
-    topic_record = crud.upsert_trending_topic(session=session, topic_data=topic_data)
-    tweets = tweets_map.get(url, [])
-    persisted_tweets = 0
-    if tweets and isinstance(tweets, list):
-        crud.replace_trending_tweets(
-            session=session, topic_id=topic_record.id, tweets_data=tweets
-        )
-        persisted_tweets = len(tweets)
-    return 1, persisted_tweets
-
-
 async def persist_scraped_batch_node(state: ScrapingGraphState) -> dict[str, Any]:
     """Persist scraped topics and tweets into PostgreSQL via CRUD upsert."""
     scraped_topics_raw = state.get("scraped_topics", [])
     scraped_topics = scraped_topics_raw if isinstance(scraped_topics_raw, list) else []
+    if not scraped_topics:
+        return {
+            "persisted_topic_count": 0,
+            "persisted_tweet_count": 0,
+            "status": "persisted",
+        }
+
     topic_tweets_map = (
         state.get("topic_tweets_map", {})
         if isinstance(state.get("topic_tweets_map"), dict)
@@ -548,64 +448,38 @@ async def persist_scraped_batch_node(state: ScrapingGraphState) -> dict[str, Any
         if isinstance(state.get("topic_summaries"), dict)
         else {}
     )
-    user_id_raw = state.get("user_id")
-
-    if not scraped_topics:
-        return {
-            "persisted_topic_count": 0,
-            "persisted_tweet_count": 0,
-            "status": "persisted",
-        }
-
-    persisted_topics = 0
-    persisted_tweets = 0
 
     try:
-        with resolve_session(session=state.get("session")) as session:
-            resolved_user_id = _resolve_user_id(user_id=user_id_raw, session=session)
-            now = datetime.now(timezone.utc)
-            errors: list[str] = []
+        (
+            persisted_topics,
+            persisted_tweets,
+            errors,
+        ) = persist_scraped_batch_records(
+            user_id_raw=state.get("user_id"),
+            session_arg=state.get("session"),
+            scraped_topics=scraped_topics,
+            topic_tweets_map=topic_tweets_map,
+            topic_summaries=topic_summaries,
+        )
 
-            for topic in scraped_topics:
-                if not topic:
-                    continue
-                try:
-                    t_count, tw_count = _persist_single_topic_record(
-                        session=session,
-                        resolved_user_id=resolved_user_id,
-                        topic=topic if isinstance(topic, dict) else {},
-                        tweets_map=topic_tweets_map,
-                        summaries=topic_summaries,
-                        now=now,
-                    )
-                    persisted_topics += t_count
-                    persisted_tweets += tw_count
-                except Exception as topic_err:
-                    logger.warning(f"Error persisting topic: {topic_err}")
-                    errors.append(str(topic_err))
-                    try:
-                        session.rollback()
-                    except Exception:
-                        pass
-
-            if errors and persisted_topics == 0:
-                return {
-                    "persisted_topic_count": 0,
-                    "persisted_tweet_count": 0,
-                    "status": "error",
-                    "error": "; ".join(errors),
-                }
-
+        if errors and persisted_topics == 0:
             return {
-                "persisted_topic_count": persisted_topics,
-                "persisted_tweet_count": persisted_tweets,
-                "status": "persisted",
+                "persisted_topic_count": 0,
+                "persisted_tweet_count": 0,
+                "status": "error",
+                "error": "; ".join(errors),
             }
-    except Exception as e:
-        logger.error(f"Error persisting scraped batch: {e}")
+
         return {
             "persisted_topic_count": persisted_topics,
             "persisted_tweet_count": persisted_tweets,
+            "status": "persisted",
+        }
+    except Exception as e:
+        logger.error(f"Error persisting scraped batch: {e}")
+        return {
+            "persisted_topic_count": 0,
+            "persisted_tweet_count": 0,
             "status": "error",
             "error": str(e),
         }
